full version

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: tkoami <tkoami@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 19:40:59 by tkoami            #+#    #+#             */
-/*   Updated: 2020/11/20 15:41:15 by tkoami           ###   ########.fr       */
+/*   Updated: 2020/11/26 14:09:33 by tkoami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,14 +16,14 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned char			*s1;
 	unsigned char			*s2;
-	int						i;
+	size_t					i;
 
 	s1 = (unsigned char*)dst;
 	s2 = (unsigned char*)src;
 	i = 0;
 	if (dst == src)
 		return (dst);
-	while (i < (int)n)
+	while (i < n)
 	{
 		s1[i] = s2[i];
 		i++;
