+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: tkoami <tkoami@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/15 16:53:45 by tkoami            #+#    #+#             */
-/*   Updated: 2020/11/25 23:14:19 by tkoami           ###   ########.fr       */
+/*   Updated: 2020/11/26 14:21:05 by tkoami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t		ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
 	size_t	srcsize;
-	int		i;
+	size_t	i;
 
 	if (dst == NULL || src == NULL)
 		return (0);
@@ -23,7 +23,7 @@ size_t		ft_strlcpy(char *dst, const char *src, size_t dstsize)
 	i = 0;
 	if (dstsize == 0)
 		return (srcsize);
-	while (src[i] != '\0' && i != (int)(dstsize - 1))
+	while (src[i] != '\0' && i != dstsize - 1)
 	{
 		dst[i] = src[i];
 		i++;
