+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: tkoami <tkoami@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/15 17:40:30 by tkoami            #+#    #+#             */
-/*   Updated: 2020/11/23 15:12:09 by tkoami           ###   ########.fr       */
+/*   Updated: 2020/11/26 14:22:34 by tkoami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*str1;
 	unsigned char	*str2;
 
@@ -23,7 +23,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 	i = 0;
 	if (n == 0)
 		return (0);
-	while (i < ((int)n - 1) && str1[i] != '\0' && str2[i] != '\0')
+	while (i < n - 1 && str1[i] != '\0' && str2[i] != '\0')
 	{
 		if (str1[i] != str2[i])
 			break ;
